requirements.txt
unidecode
py-cpuinfo
gpustat
psutil
rich